requirements.txt
Flask
SQLAlchemy
psycopg2
bcrypt
PyJWT
fpdf
